Take into account missing images or missing matches

+++ tagger/scripts/fetch_sift.py
@@ -8,12 +8,17 @@ OUTPUT = '../data/sift/'
 
 def download_photo(photo_id, user_id):
     url = 'https://www.flickr.com/photos/{}/{}'.format(user_id, photo_id)
-    html = urllib2.urlopen(url).read()
-    img_url = re.findall(r'(?:https?://)?farm[^":]+_o\.(?:jpg|gif|png)', html)[0].replace('\/', '/')
-    filename = OUTPUT + photo_id + '.jpg'
+    try:
+        html = urllib2.urlopen(url).read()
+    except urllib2.HTTPError, err:
+        return
+    img_urls = re.findall(r'(?:https?://)?farm[^":]+_o\.(?:jpg|gif|png)', html)
+    if len(img_urls) > 0:
+        img_url = img_urls[0].replace('\/', '/')
+        filename = OUTPUT + photo_id + '.jpg'
 
-    with open(filename, "wb") as fp:
-        fp.write(urllib2.urlopen('https://' + img_url).read())
+        with open(filename, "wb") as fp:
+            fp.write(urllib2.urlopen('https://' + img_url).read())
 
 
 def fetch_sift(photo_id, user_id):
@@ -21,8 +26,9 @@ def fetch_sift(photo_id, user_id):
     if os.path.isfile(filename + '.SIFT'):
         return
     download_photo(photo_id, user_id)
-    extract_sift(filename + '.jpg')
-    os.remove(filename + '.jpg')
+    if os.path.isfile(filename + '.jpg'):
+        extract_sift(filename + '.jpg')
+        os.remove(filename + '.jpg')
 
 # Download images and extract SIFT descriptors of the training set
 with open(INPUT_TRAIN) as csvfile:
@@ -34,3 +40,5 @@ with open(INPUT_TRAIN) as csvfile:
         print 'Fetching file #{}'.format(num)
         fetch_sift(photo_id, user_id)
         num += 1
+        if num == 5001:
+            break
